Add support for ignoreList property
Sourcemaps can now specify an `ignoreList`, a list of indexes into sources for which those sources should be shown as ignored in dev tools, for example as collapsed stack frames in program stacks.

This adds the list to the SourceMap structure but only includes it in serialized output when it’s non-zero. Open to feedback.

Added a test to exercise this through flattening indexed source maps into a regular map with updated source positions.